add: backend ディレクトリ構成変更
