+++ backend/utils/log.py
@@ -1,0 +1,8 @@
+import logging
+
+# ロギングの設定を強化
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
